requirements.txt
streamlit>=1.27.0
openai>=1.0.0
python-dotenv